+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Added Enhanced Auto Crafter
 * Added "Smart-Filling" mode to Cargo Input nodes
 * Added "Netherite Ingot -> Netherite Block" recipe to Electric Press
+* Added Armor Forge Auto Crafter
 
 #### Changes
 * Removed all functionality from the old Automated Crafting Chamber